Replace image of H1gNOeHKPS
The existing image is a pure-text crop of the experimental setup which
doesn't reflect the content of the paper.

This PR replaces that image with Figure 1 on page 2, which is a
visualization of the primary contribution. This image also works well
with this spotlight awarded paper being in the network architecture
category.